+++ finance/lending/anchor/programs/lending/src/instructions/admin/mod.rs
@@ -1,8 +1,10 @@
+pub mod collect_protocol_fees;
 pub mod init_lending_market;
 pub mod init_reserve;
 pub mod set_price;
 pub mod update_reserve_config;
 
+pub use collect_protocol_fees::*;
 pub use init_lending_market::*;
 pub use init_reserve::*;
 pub use set_price::*;